初期セットアップの部分のメソッドの実装とブロックの暫定追加など。
＜詳細＞
【GingaCore】GMathHelperクラスを作成。
【GingaCore】MQO関連のメソッドを修正（float→doubleなど）
【GTC】ブロック、タブ、ログなどのコードを追加。

+++ src/main/java/jp/gingarenpo/api/helper/GPosHelper.java
@@ -68,7 +68,7 @@ public class GPosHelper {
 	 * @param pos 調べたい数値。floatでいくらでも指定できます。
 	 * @return 全てが一致していればtrue、1つでも異なればfalse
 	 */
-	public static boolean areSameNumber(float ...pos) {
+	public static boolean areSameNumber(double ...pos) {
 		// 全ての数値が一致していたらtrue
 		for (int i = 0; i < pos.length - 1; i++) {
 			if (pos[i] != pos[i+1]) return false; // 1個でも違えば残念ながら不一致
